NEW cumul temps des taches dans une tache si c'est la même workstation

+++ class/ordre_fabrication_asset.class.php
@@ -2834,14 +2834,26 @@ class TAssetWorkstationOF extends TObjetStd{
 		$projectTask->fetch($this->fk_project_task);
 		$projectTask->fk_project = $OF->fk_project;
 
-		if($projectTask->planned_workload<=0)  $projectTask->planned_workload = $this->nb_hour*3600;
+        if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && !empty($OF->from_create)) $projectTask->planned_workload += $this->nb_hour * 3600; // On cumul le temps dans la tache
+        else if($projectTask->planned_workload <= 0) $projectTask->planned_workload = $this->nb_hour * 3600;
 
-		if(empty($conf->gantt->enabled)) {
-			$projectTask->date_start = strtotime(' +'.(int)$this->nb_days_before_beginning.'days',$OF->date_lancement);
-			$projectTask->date_end = $OF->date_besoin;
-			if($projectTask->date_end<$projectTask->date_start)$projectTask->date_end = $projectTask->date_start;
+        if(empty($conf->gantt->enabled)) {
+            if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && !empty($OF->from_create)) {
 
-		}
+                //On prend la date la plus petite
+                if($projectTask->date_start > $OF->date_lancement) $projectTask->date_start = strtotime(' +' . (int)$this->nb_days_before_beginning . 'days', $OF->date_lancement);
+
+                //On prend la date la plus grande
+                if($projectTask->date_end < $OF->date_besoin) $projectTask->date_end = $OF->date_besoin;
+
+                if($projectTask->date_end < $projectTask->date_start) $projectTask->date_end = $projectTask->date_start;
+            }
+            else {
+                $projectTask->date_start = strtotime(' +' . (int)$this->nb_days_before_beginning . 'days', $OF->date_lancement);
+                $projectTask->date_end = $OF->date_besoin;
+                if($projectTask->date_end < $projectTask->date_start) $projectTask->date_end = $projectTask->date_start;
+            }
+        }
 
 		$projectTask->update($user);
 
@@ -2891,7 +2903,27 @@ class TAssetWorkstationOF extends TObjetStd{
 
 		$action = '';
 
-		if ($of->fk_project > 0 && $this->fk_project_task == 0) $action = 'createTask';
+		if ($of->fk_project > 0 && $this->fk_project_task == 0){
+
+		    $action = 'createTask';
+
+            if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK)){
+
+                $taskstatic = new Task($db);
+                $TTask = $taskstatic->getTasksArray(null, null, $of->fk_project);
+                if(!empty($TTask)) {
+                    foreach($TTask as $task) {
+                        $task->fetch_optionals();
+                        if(!empty($task->array_options['options_fk_workstation']) && $this->fk_asset_workstation == $task->array_options['options_fk_workstation']){
+                            $action = 'updateTask';
+                            $this->fk_project_task=$task->id;
+                            $of->from_create=1;
+                        }
+
+                    }
+                }
+            }
+        }
 		elseif ($of->fk_project > 0 && $this->fk_project_task > 0) $action = 'updateTask';
 		elseif ($of->fk_project == 0 && $this->fk_project_task > 0) $action = 'deleteTask';
 
@@ -2904,7 +2936,7 @@ class TAssetWorkstationOF extends TObjetStd{
 				$this->updateTask($PDOdb, $db, $conf, $user, $of);
 				break;
 			case 'deleteTask':
-				$this->deleteTask($db, $conf, $user);
+                if(empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) || !empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && $this->isLastLink())$this->deleteTask($db, $conf, $user);
 				break;
 			default:
 				break;
@@ -3033,7 +3065,7 @@ class TAssetWorkstationOF extends TObjetStd{
 			if($projectTask->fetch($this->fk_project_task) > 0) {
 				// Suppression des occurences qui définissent cette tâches en tant que parente
 				$db->query('UPDATE '.MAIN_DB_PREFIX.'projet_task SET fk_task_parent = 0 WHERE fk_task_parent = '.$projectTask->id);
-				$projectTask->delete($user);
+                if(empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) || !empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && $this->isLastLink())$projectTask->delete($user);
 			}
 		}
 
@@ -3190,6 +3222,18 @@ class TAssetWorkstationOF extends TObjetStd{
 		$projectTask->progress = $progress;
 		$projectTask->update($db);
 	}
+
+	function isLastLink(){
+	    global $db;
+
+	    $sql = "SELECT rowid FROM ".MAIN_DB_PREFIX."asset_workstation_of WHERE fk_project_task=".$this->fk_project_task;
+	    $resql = $db->query($sql);
+	    $rows = $db->num_rows($resql);
+
+	    if(!empty($rows)) return false;
+
+	    return true;
+    }
 
 }
